+++ src/plugins/relay/api/relay-api-protocol.c
@@ -797,7 +797,7 @@ RELAY_API_PROTOCOL_CALLBACK(input)
             {
                 relay_api_msg_send_error_json (
                     client,
-                    RELAY_HTTP_400_BAD_REQUEST, NULL,
+                    RELAY_HTTP_404_NOT_FOUND, NULL,
                     "Buffer \"%s\" not found",
                     ptr_buffer_name);
                 cJSON_Delete (json_body);
@@ -819,7 +819,7 @@ RELAY_API_PROTOCOL_CALLBACK(input)
                 {
                     relay_api_msg_send_error_json (
                         client,
-                        RELAY_HTTP_400_BAD_REQUEST, NULL,
+                        RELAY_HTTP_404_NOT_FOUND, NULL,
                         "Buffer \"%lld\" not found",
                         (long long)cJSON_GetNumberValue (json_buffer_id));
                     cJSON_Delete (json_body);
@@ -924,7 +924,7 @@ RELAY_API_PROTOCOL_CALLBACK(completion)
             {
                 relay_api_msg_send_error_json (
                     client,
-                    RELAY_HTTP_400_BAD_REQUEST, NULL,
+                    RELAY_HTTP_404_NOT_FOUND, NULL,
                     "Buffer \"%s\" not found",
                     ptr_buffer_name);
                 cJSON_Delete (json_body);
@@ -946,7 +946,7 @@ RELAY_API_PROTOCOL_CALLBACK(completion)
                 {
                     relay_api_msg_send_error_json (
                         client,
-                        RELAY_HTTP_400_BAD_REQUEST, NULL,
+                        RELAY_HTTP_404_NOT_FOUND, NULL,
                         "Buffer \"%lld\" not found",
                         (long long)cJSON_GetNumberValue (json_buffer_id));
                     cJSON_Delete (json_body);